% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcoxp-package.R
\docType{package}
\name{abcoxp-package}
\alias{abcoxp}
\alias{abcoxp-package}
\title{abcoxp: Implement Approximate Bayesian Inference for Cox Proportional Hazard Model with Partial Likelihood}
\description{
This package implements the approximate Bayesian inference method described in the paper "Bayesian Inference for Cox Proportional Hazard Models with Partial Likelihoods, Semi-Parametric Covariate Effects, and Correlated Observations" by Zhang et al. (2022).
}
\keyword{internal}
