% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_construction.R
\name{abcoxp_fit}
\alias{abcoxp_fit}
\title{Do the Approximate Bayesian Inference for the specified Cox
Proportional Hazard Model.}
\usage{
abcoxp_fit(
  data,
  times,
  cens,
  fixed = NULL,
  frailty = NULL,
  RW2 = NULL,
  fixed_control = list(betaprec = 0.001),
  frailty_control = list(alpha = 0.5, u = 1),
  RW2_control = list(alpha = 0.5, u = 1, r = 50),
  Inference_control = list(aghq_k = 4),
  diag_noise = 1e-04
)
}
\arguments{
\item{data}{A dataframe that should contain all the variables listed as times, cens, fixed,
frailty and RW2.}

\item{times}{A single string of variable name to be used as survival times.}

\item{cens}{A single string of variable name to be used as censoring indicators}

\item{fixed}{A vector of strings of covariate names to be used as linear
fixed effects in the model. All covariates are assumed to be numeric.}

\item{frailty}{A single string of covariate name to be used as the group for
frailty in the model, default is NULL meaning no frailty.}

\item{RW2}{A single string of covariate name to be used as the RW2 smoothing
covariate, default is NULL meaning no RW2 smoothing. The covariate is assumed to be
numeric.}

\item{fixed_control}{A list that specifies the prior precision of all the fixed
effect parameters. Default being list(betaprec = .001).}

\item{frailty_control}{A list that specifies the PC prior for the frailty SD
parameters. Default being list(alpha = 0.5, u = 1)}

\item{RW2_control}{A list that specifies the PC prior for the RW2 SD
parameters and number of knots used. Default being list(alpha = 0.5, u = 1, r = 50)}

\item{Inference_control}{A list that specifies the setting of the AGHQ inference, including
the number of grid points K. Default being list(aghq_k = 4).}

\item{diag_noise}{The small amount of diagonal adjustment to add when RW2 exists in the model.
The default value is diag_noise = 0.0001.}
}
\value{
A list that contains the fitted AGHQ object of the model and other components.
}
\description{
Do the Approximate Bayesian Inference for the specified Cox
Proportional Hazard Model.
}
