% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{abcoxp_sampling}
\alias{abcoxp_sampling}
\title{Implements the Approximate Bayesian Inference for the specified Cox
Proportional Hazard Model with Analysis of Posterior.}
\usage{
abcoxp_sampling(fitted_model, nsamps)
}
\arguments{
\item{fitted_model}{A list from the output of \code{abcoxp_fit}.}

\item{nsamps}{An integer value of the number of samples.}
}
\value{
A list that contains the posterior samples.
}
\description{
Implements the Approximate Bayesian Inference for the specified Cox
Proportional Hazard Model with Analysis of Posterior.
}
