% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_construction.R
\name{RW2_design_comp}
\alias{RW2_design_comp}
\title{Compute the design matrix of the RW2 part:}
\usage{
RW2_design_comp(x, r)
}
\arguments{
\item{x}{A vector of covariate values.}

\item{r}{An integer number of knots.}
}
\value{
A sparse matrix denotes the design matrix of the RW2 component.
}
\description{
Compute the design matrix of the RW2 part:
}
