% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_construction.R
\name{RW2_prec_comp}
\alias{RW2_prec_comp}
\title{Compute the precision matrix of the RW2 part:}
\usage{
RW2_prec_comp(x, r, diag_noise)
}
\arguments{
\item{x}{A vector of covariate values.}

\item{r}{An integer number of knots.}

\item{diag_noise}{A positive number denotes the diagonal adjustment.}
}
\value{
A sparse matrix denotes the precision matrix of the RW2 component.
}
\description{
Compute the precision matrix of the RW2 part:
}
